🔨 refactor: Home 페이지 컴포넌트, 타입 분리
