+++ src/app/home/page.tsx
@@ -1,117 +1,22 @@
 "use client";
-import React, { useEffect, useState } from "react";
-import Image from "next/image";
-import { HomeMain, MainBanner } from "@/styles/homeStyle";
-import bannerImg from "../../../public/assets/images/banner.png";
-import ProductSection, {
-  GroupOrderList,
-} from "@/components/homeSection/productSection";
-import GroupSection from "@/components/homeSection/groupSection";
-import MarketTab, { MarketItem } from "@/components/postList/marketTab";
-import GroupPurchaseTab from "@/components/postList/groupPurchaseTab";
-import { useRouter } from "next/navigation";
-import useApi from "@/hooks/common/useApi";
+import React from "react";
+import { HomeMain } from "@/styles/homeStyle";
+import useHomeData from "@/hooks/pages/home/useHomeData";
+import HomeBanner from "@/components/pages/home/homeBanner";
+import HomeTab from "@/components/pages/home/homeTab";
 
-export interface HomeList {
-  post_id: number;
-  user_id: string;
-  board_type: string;
-  title: string;
-  content: string;
-  file_group_id: string;
-  view_count: number;
-  insert_dt: string;
-  update_dt: string | null;
-  post_status: string;
-}
-
-export interface HomeData {
-  groupOrderList: GroupOrderList[];
-  marketList: MarketItem[];
-}
-
-const Home: React.FC = () => {
-  const [data, setData] = useState<HomeData | null>(null);
-  const router = useRouter();
-
-  // 탭 관리
-  const [tabStatus, setTabStatus] = useState<string>("Home");
-
-  useEffect(() => {
-    const fetchHomeData = async () => {
-      try {
-        const res = await useApi({
-          method: "get",
-          url: "home/list",
-          params: { status: "InProgress" },
-        });
-        setData(res);
-      } catch (error) {
-        console.error("홈 데이터 불러오기 실패:", error);
-      }
-    };
-
-    fetchHomeData();
-  }, []);
-
-  const setActiveClass = (status: string) => {
-    return tabStatus === status ? "active" : "";
-  };
-
-  const renderTabContent = () => {
-    switch (tabStatus) {
-      case "Home":
-        return (
-          <>
-            <ProductSection marketList={data?.marketList || []} />
-            <GroupSection groupOrderList={data?.groupOrderList || []} />
-          </>
-        );
-      case "Market":
-        return <MarketTab api={"hometab"} />;
-      case "GroupPurchase":
-        return <GroupPurchaseTab api={"hometab"} />;
-      default:
-        return null;
-    }
-  };
+export default function Home() {
+  // 홈 데이터 리스트 불러오기
+  const { data } = useHomeData();
 
   return (
     <HomeMain>
       <h1 className="a11y-hidden">홈</h1>
-      <MainBanner>
-        <Image
-          onClick={() => router.push("/groupDetail/InProgress/11")}
-          priority
-          src={bannerImg}
-          alt="배너 이미지"
-        />
-      </MainBanner>
-      <nav>
-        <ul className="menu_tab">
-          <li
-            className={`btn_tab ${setActiveClass("Home")}`}
-            onClick={() => setTabStatus("Home")}
-          >
-            메인
-          </li>
-          <li
-            className={`btn_tab ${setActiveClass("Market")}`}
-            onClick={() => setTabStatus("Market")}
-          >
-            판매 / 구매 / 교환
-          </li>
-          <li
-            className={`btn_tab ${setActiveClass("GroupPurchase")}`}
-            onClick={() => setTabStatus("GroupPurchase")}
-          >
-            공동구매
-          </li>
-        </ul>
-      </nav>
-      {renderTabContent()}
+      <HomeBanner />
+      <HomeTab
+        marketList={data?.marketList}
+        groupOrderList={data?.groupOrderList}
+      />
     </HomeMain>
   );
-};
-
-export default Home;
+}
